Update backup/restore procedures
in Kiwi TCMS 10.3 we're not going to have the mysql/psql binaries
so backing up and restore must be done using DB specific tools from
the kiwi_db container instead of using ./manage.py dbshell on the
kiwi_web container.

+++ content/2018-07-30-docker-backup.markdown
@@ -12,56 +12,60 @@ When you start Kiwi TCMS by running `docker-compose up`
 it will automatically create 2 volumes: `kiwi_db_data` and `kiwi_uploads`.
 This blog post will outline how to backup these docker volumes.
 
-Backing up the database
------------------------
-
-Kiwi TCMS is a Django application and the `manage.py` command provides an easy way
-to dump and load the database contents. To export all contents on your docker host
-execute:
-
-    docker exec -it kiwi_web /Kiwi/manage.py dumpdata --all --indent 2 > database.json
-
-This will create a file named `database.json` in the current directory, outside of the
-running container!
-
-You can restore the database contents by using the following commands:
-
-    # delete data from all tables
-    docker exec -it kiwi_web /bin/bash -c '/Kiwi/manage.py sqlflush | /Kiwi/manage.py dbshell'
-    # then reload the existing data
-    cat database.json | docker exec -i kiwi_web /Kiwi/manage.py loaddata --format json -
-
-**NOTE:** depending on your scenario you may want to remove the existing volume
-(`docker-compose down && docker volume rm kiwi_db_data`) and re-create the
-DB schema (`/Kiwi/manage.py migrate`) before restoring the contents!
+**Note:** in the instructions below `kiwi_db` is the container name and `kiwi` is the
+database name used inside the `docker-compose.yml` file!
 
 
-**WARNING**: the above steps are applicable to Kiwi TCMS 5.1 or above. On earlier
-versions `manage.py` will fail due to various issues.
+MariaDB/MySQL database
+----------------------
+
+To export all contents from the docker container execute the command:
+
+    docker exec -u 0 -i kiwi_db mysqldump kiwi > backup.sql
+
+This will create a file named `backup.sql` in the current directory, outside of the running container!
+
+You can restore the database contents by using the following command:
+
+    cat backup.sql | docker exec -u 0 -i kiwi_db mysql -v kiwi
+
+**Notes:**
+
+1. The commands above using `-u 0` are executed with `root` privileges inside the
+   container. In this way you don't need to type-in the database password!
+2. Depending on your scenario you may want to remove the existing volume
+   (`docker-compose down && docker volume rm kiwi_db_data`) before restoring the database!
 
 
-Backing up multi-tenant database
---------------------------------
+Postgres database
+-----------------
 
-The [kiwitcms-tenant](https://github.com/kiwitcms/tenants) add-on depends on the
-PostgreSQL database. It will create multiple DB schemas, one per tenant. To backup
-all tenants use the following command:
+To export all contents from the docker container execute the command:
 
-    docker exec -i kiwi_db /bin/bash -c 'pg_dump --dbname=kiwi -F c' > backup.bak
+    docker exec -i kiwi_db pg_dump --dbname=kiwi -F c > backup.bak
 
-This will create a file in the PostgreSQL custom database dump format which
-contains all data and schema definitions. That is a binary file which can be read
-with the `pg_restore` command.
+This will create a file named `backup.bak` in the current directory, outside of the running container.
+This is a PostgreSQL custom database dump format which contains all data and schema definitions.
+That is a binary file which can be read with the `pg_restore` command!
 
-To [drop and] restore the entire multi-tenant database:
+To drop and restore the entire database execute:
 
-    docker exec -i kiwi_db /bin/bash -c 'psql -c "DROP DATABASE IF EXISTS kiwi;"'
-    cat backup.bak | docker exec -i kiwi_db /bin/bash -c 'pg_restore --dbname=template1 -vcC'
+    docker exec -i kiwi_db psql -c "DROP DATABASE IF EXISTS kiwi;"
+    cat backup.bak | docker exec -i kiwi_db pg_restore --dbname=template1 -vcC
 
 
-To [drop and] restore an individual tenant:
+Multi-tenant database
+---------------------
 
-    docker exec -it kiwi_web /Kiwi/manage.py dbshell
+The [kiwitcms-tenant](https://github.com/kiwitcms/tenants) add-on and/or
+[Kiwi TCMS Enterprise](https://github.com/kiwitcms/enterprise) work only on
+Postgres! Each tenant (aka name-space) uses a separate database schema.
+The first schema name is `public`.
+
+The backup and restore instructions shown above operate on all tenants together!
+If you want to [drop and] restore an individual tenant then use the commands:
+
+    docker exec -it kiwi_db psql --dbname=kiwi
     
     kiwi=> DROP SCHEMA $tenant_name CASCADE;
     ....
@@ -70,11 +74,7 @@ To [drop and] restore an individual tenant:
     CREATE SCHEMA
     kiwi=>Ctrl+D
     
-    cat backup.bak | docker exec -i kiwi_db /bin/bash -c 'pg_restore --dbname=kiwi -v --schema $tenant_name'
-
-
-**WARNING:** `sqlflush | dbshell` will not work when you have multiple DB schemas so you must use
-the PostgreSQL database shell to manipulate the contents of the database!
+    cat backup.bak | docker exec -i kiwi_db pg_restore --dbname=kiwi -v --schema $tenant_name
 
 
 Backing up file uploads
@@ -84,15 +84,17 @@ Uploaded files can easily be backed up with:
 
     docker exec -i kiwi_web /bin/tar -cP /Kiwi/uploads > uploads.tar
 
-and then restored:
+and then restored with:
 
     cat uploads.tar | docker exec -i kiwi_web /bin/tar -x
 
 You may also try the `rsync` command but be aware that it is not installed
 by default!
 
-The same approach may be used to backup `/var/lib/mysql/` from the `kiwi_db`
-container.
+**Note:**
+the same approach may be used to backup `/var/lib/mysql/` or `/var/lib/pgsql/data`
+from the `kiwi_db` container.
+
 
 Backing up multi-tenant uploads
 -------------------------------
